allow hue values larger than byte

+++ tplink-smartdevices/Data/BulbHSV.cs
@@ -2,8 +2,8 @@
 {
     public class BulbHSV
     {
-        public byte Hue { get; set; }
-        public byte Saturation { get; set; }
-        public byte Value { get; set; }
+        public int Hue { get; set; }
+        public int Saturation { get; set; }
+        public int Value { get; set; }
     }
 }
